conditional constraints review fixes
